ODK Briefcase R API

ODK Briefcase is a Java application for fetching and pushing OpenDataKit (ODK) forms and their contents. It can be used to gather and aggregate data from the mobile client ODK Collect when there is no internet connectivity, when there is no ODK Aggregate Server set-up or when an ODK Aggregate Server is not preferred. Current production version of ODK Briefcase is v1.8.0 and an executable .jar file can be downloaded from the OpenDataKit GitHub repo.

Details

As of ODK Briefcase version 1.4.4, a scriptable command line interface (CLI) to the Java application has been available. This package provides an R interface to ODK Briefcase via the available CLI to pull forms from a remote ODK Aggregate Server or from a local ODK folder /odk collected from mobile clients. This package requires Java 8 installed. Java 8 can be downloaded here.

See also