Sample dataset from an impact evaluation study of a mother and child nutrition programme in Kassala State, Sudan. This dataset contains cluster level data from the survey.

Sample dataset from an impact evaluation study of a mother and child nutrition programme in Kassala State, Sudan. This dataset contains cluster level data from the survey.

sampleData1

Format

A data frame with 16 columns and 50 rows:

admin.admin1.adm1

Date of survey

admin.admin1.adm2

Enumerator ID

admin.enameA

Name of enumerator (Arabic)

admin.ename

Name of enumerator (English)

admin.admin2.adm3

Survey round number

admin.admin2.adm4

Study area / cluster

location.loc1

Village ID

location.loc1a

Village ID - other

location.loc2

Is village a replacement village

location.loc3

Village population

location.loc4

GPS coordinates

hh1

Household ID

hh2

Number of women aged 15-49 in household

wcount_count

Current woman respondent's ID

KEY

Parent data identifier

PARENT_KEY

Child data identifier

Examples

sampleData1

#>    admin.admin1.adm1 admin.admin1.adm2              admin.ename
#> 1         2016-12-24                 4  Eman Abdelaleem Zakaria
#> 2         2016-12-24                 7        Hafsa Ali Mohamed
#> 3         2016-12-24                 5          Fatima Abdallah
#> 4         2016-12-24                 1            Amani Elkheir
#> 5         2016-12-24                20             Zahra Younis
#> 6         2016-12-24                13 Mahasin Mohamed Elhassan
#> 7         2016-12-24                11             Khadija Adam
#> 8         2016-12-24                18          Thuraya Mahmoud
#> 9         2016-12-24                 5          Fatima Abdallah
#> 10        2016-12-24                 7        Hafsa Ali Mohamed
#> 11        2016-12-24                19            Thuraya Osman
#> 12        2016-12-24                 7        Hafsa Ali Mohamed
#> 13        2016-12-24                13 Mahasin Mohamed Elhassan
#> 14        2016-12-24                 1            Amani Elkheir
#> 15        2016-12-24                18          Thuraya Mahmoud
#> 16        2016-12-24                20             Zahra Younis
#> 17        2016-12-24                 5          Fatima Abdallah
#> 18        2016-12-24                11             Khadija Adam
#> 19        2016-12-24                19            Thuraya Osman
#> 20        2016-12-24                20             Zahra Younis
#> 21        2016-12-24                 4  Eman Abdelaleem Zakaria
#> 22        2016-12-24                 2             Amira Abdeen
#> 23        2016-12-24                 5          Fatima Abdallah
#> 24        2016-12-24                17      Sabah Salih Ibrahim
#> 25        2016-12-24                 6            Fatima Alamin
#> 26        2016-12-24                 7        Hafsa Ali Mohamed
#> 27        2016-12-24                18          Thuraya Mahmoud
#> 28        2016-12-24                 9      Hayat Mohamed Idris
#> 29        2016-12-24                13 Mahasin Mohamed Elhassan
#> 30        2016-12-24                 1            Amani Elkheir
#> 31        2016-12-24                15           Omaima Homeida
#> 32        2016-12-24                14       Manihil Fadl Alsid
#> 33        2016-12-24                 7        Hafsa Ali Mohamed
#> 34        2016-12-24                 2             Amira Abdeen
#> 35        2016-12-24                11             Khadija Adam
#> 36        2016-12-24                 5          Fatima Abdallah
#> 37        2016-12-24                17      Sabah Salih Ibrahim
#> 38        2016-12-24                 6            Fatima Alamin
#> 39        2016-12-24                 4  Eman Abdelaleem Zakaria
#> 40        2016-12-24                18          Thuraya Mahmoud
#> 41        2016-12-24                 9      Hayat Mohamed Idris
#> 42        2016-12-24                13 Mahasin Mohamed Elhassan
#> 43        2016-12-24                 3       Awadia Mohamed Ali
#> 44        2016-12-24                 5          Fatima Abdallah
#> 45        2016-12-24                 7        Hafsa Ali Mohamed
#> 46        2016-12-24                18          Thuraya Mahmoud
#> 47        2016-12-24                 5          Fatima Abdallah
#> 48        2016-12-24                20             Zahra Younis
#> 49        2016-12-24                 2             Amira Abdeen
#> 50        2016-12-24                 9      Hayat Mohamed Idris
#>        admin.enameA admin.admin2.adm3 admin.admin2.adm4 location.loc1
#> 1  _____ ___ ______                 4                 6             6
#> 2     ____ ___ ____                 4                 4             1
#> 3    _____ ___ ____                 3                 6             6
#> 4       _____ _____                 4                 5             2
#> 5         ____ ____                 4                 7            17
#> 6        _____ ____                 4                 3            10
#> 7         _____ ___                 4                 3            10
#> 8         ____ ____                 4                 6             6
#> 9    _____ ___ ____                 4                 6             6
#> 10    ____ ___ ____                 4                 4             1
#> 11       ____ _____                 4                 5             2
#> 12    ____ ___ ____                 4                 4             1
#> 13       _____ ____                 4                 3            10
#> 14      _____ _____                 4                 5             2
#> 15        ____ ____                 4                 6             6
#> 16        ____ ____                 4                 7            17
#> 17   _____ ___ ____                 3                 6             6
#> 18        _____ ___                 4                 3            10
#> 19       ____ _____                 4                 5             2
#> 20        ____ ____                 4                 7            17
#> 21 _____ ___ ______                 4                 6             6
#> 22     _____ ______                 4                 7            17
#> 23   _____ ___ ____                 4                 6             6
#> 24        ____ ____                 4                 5             2
#> 25     _____ ______                 4                 3            10
#> 26    ____ ___ ____                 4                 4             1
#> 27        ____ ____                 4                 6             6
#> 28        ____ ____                 4                 1             3
#> 29       _____ ____                 4                 3            10
#> 30      _____ _____                 4                 5             2
#> 31      _____ _____                 4                 1             3
#> 32  _____ ___ _____                 4                 4             1
#> 33    ____ ___ ____                 4                 4             1
#> 34     _____ ______                 4                 7            17
#> 35        _____ ___                 4                 3            10
#> 36   _____ ___ ____                 4                 6             6
#> 37        ____ ____                 4                 1             3
#> 38     _____ ______                 4                 3            10
#> 39 _____ ___ ______                 4                 6             6
#> 40        ____ ____                 4                 6             6
#> 41        ____ ____                 4                 1             2
#> 42       _____ ____                 4                 3            10
#> 43       _____ ____                 4                 4             1
#> 44   _____ ___ ____                 4                 6             6
#> 45    ____ ___ ____                 4                 4             1
#> 46        ____ ____                 4                 6             6
#> 47   _____ ___ ____                 4                 6             6
#> 48        ____ ____                 4                 7            17
#> 49     _____ ______                 4                 7            17
#> 50        ____ ____                 4                 1             2
#>    location.loc1a location.loc2 location.loc3
#> 1                             2           300
#> 2                             2          1500
#> 3                             2           300
#> 4                             2          1000
#> 5                             2          1000
#> 6                             2           150
#> 7                             2           150
#> 8                             2           300
#> 9                             2           300
#> 10                            2          1500
#> 11                            2          1000
#> 12                            2          1500
#> 13                            2           150
#> 14                            2          1000
#> 15                            2           300
#> 16                            2          1000
#> 17                            2           300
#> 18                            2           150
#> 19                            2          1000
#> 20                            2          1000
#> 21                            2           300
#> 22                            2          1000
#> 23                            2           300
#> 24                            2          1000
#> 25                            2           150
#> 26                            2          1500
#> 27                            2           300
#> 28                            2           100
#> 29                            2           150
#> 30                            2          1000
#> 31                            2           100
#> 32                            2          1500
#> 33                            2          1500
#> 34                            2          1000
#> 35                            2           150
#> 36                            2           300
#> 37                            2           100
#> 38                            2           150
#> 39                            2           300
#> 40                            2           300
#> 41                            2           100
#> 42                            2           150
#> 43                            2          1500
#> 44                            2           300
#> 45                            2          1500
#> 46                            2           300
#> 47                            2           300
#> 48                            2          1000
#> 49                            2          1000
#> 50                            2           100
#>                                                    location.loc4 hh1 hh2
#> 1  15.340707055739859 36.409332161777634 489.62804045528173 20.0  10   1
#> 2    15.243085888579019 35.50734621989882 470.2089592535049 25.0   2   1
#> 3                                                                 18   1
#> 4    14.863367945089205 35.866666603118375 -1.9034378901124 15.0   5   1
#> 5    15.468007832681407 36.43122575900205 521.5972869871184 25.0  11   1
#> 6    15.858674407154577 36.37231720853207 454.6278015691787 10.0   6   1
#> 7                                                                  7   1
#> 8     15.344821969653859 36.4123056904053 524.6815485022962 20.0  19   1
#> 9   15.342416367468351 36.409081857962754 509.3362386887893 15.0  19   1
#> 10   15.242720015920215 35.50733519201448 461.8070472544059 10.0   3   1
#> 11   14.863197175573642 35.86928686765128 462.4303112337366 10.0   5   1
#> 12  15.242760667735546 35.507351704810645 468.1484348261729 20.0   4   1
#> 13                                                                 7   1
#> 14                                                                 6   1
#> 15   15.344836542984442 36.412303807482466 528.715526827611 25.0  20   1
#> 16                                                                12   1
#> 17                                                                20   1
#> 18    15.859804355349041 36.3697683675118 467.4077677382156 10.0   8   1
#> 19   14.86303281506242 35.86851605079117 485.36729476600885 15.0   6   1
#> 20                                                                13   1
#> 21  15.340626004927747 36.409704627473516 531.3647508798167 15.0  11   1
#> 22   15.468108546033292 36.43084176907657 529.8765007127076 20.0  10   1
#> 23   15.341934954061003 36.40920065801615 509.6119866305962 20.0  21   1
#> 24                                                                 6   1
#> 25 15.857047611120107 36.367080908743965 457.56518397387117 25.0  10   1
#> 26 15.242564793065485 35.507391095599594 459.71703636739403 15.0   5   1
#> 27  15.345046529712212 36.412365115620794 529.1299370490015 10.0  21   1
#> 28   15.508623605231321 35.71921247885054 393.2120068129152 10.0   7   1
#> 29                                                                 8   1
#> 30  14.863073715960105 35.868534599770236 475.1144825257361 25.0   7   1
#> 31  15.507955124315897 35.71994656880656 435.78340286947787 15.0   1   1
#> 32      15.243935478477816 35.5072483748536 462.23733238969 20.0   2   1
#> 33  15.243544853526092 35.504358697667676 318.2708964664489 10.0   6   1
#> 34  15.468348778718005 36.430295749140896 527.3553353380412 25.0  11   1
#> 35  15.858663032550746 36.37233865660788 453.68353387061507 15.0   9   1
#> 36    15.34195942957224 36.40936281884173 542.9970574751496 20.0  22   1
#> 37     15.50791103507599 35.71998905828642 443.707229337655 15.0   1   1
#> 38                                                                11   1
#> 39  15.340625957958748 36.409617272319956 522.9894889993593 15.0  12   1
#> 40  15.345026304693446 36.411030820483106 511.9512688117102 15.0  22   1
#> 41  15.507898530115408 35.72003764021008 437.78598945215344 10.0   2   1
#> 42                                                                 9   1
#> 43   15.243843041379714 35.50728506878729 464.1237193001434 20.0   1   1
#> 44                                                                23   1
#> 45   15.241833111246876 35.50801737078307 461.0185168450698 15.0   7   1
#> 46                                                                23   1
#> 47    15.34193913638263 36.40920103182395 515.2498174784705 15.0  24   1
#> 48   15.468200337263658 36.43041438343925 517.0020088292658 20.0  13   1
#> 49    15.4686049946019 36.429610500645374 512.2657940238714 20.0  12   1
#> 50                                                                 3   1
#>    wcount_count KEY PARENT_KEY
#> 1             1   1         -1
#> 2             1   2         -1
#> 3             1   3         -1
#> 4             1   4         -1
#> 5             1   5         -1
#> 6             1   6         -1
#> 7             1   7         -1
#> 8             1   8         -1
#> 9             1   9         -1
#> 10            1  10         -1
#> 11            1  11         -1
#> 12            1  12         -1
#> 13            1  13         -1
#> 14            1  14         -1
#> 15            1  15         -1
#> 16            1  16         -1
#> 17            1  17         -1
#> 18            1  18         -1
#> 19            1  19         -1
#> 20            1  20         -1
#> 21            1  21         -1
#> 22            1  22         -1
#> 23            1  23         -1
#> 24            1  24         -1
#> 25            1  25         -1
#> 26            1  26         -1
#> 27            1  27         -1
#> 28            1  28         -1
#> 29            1  29         -1
#> 30            1  30         -1
#> 31            1  31         -1
#> 32            1  32         -1
#> 33            1  33         -1
#> 34            1  34         -1
#> 35            1  35         -1
#> 36            1  36         -1
#> 37            1  37         -1
#> 38            1  38         -1
#> 39            1  39         -1
#> 40            1  40         -1
#> 41            1  41         -1
#> 42            1  42         -1
#> 43            1  43         -1
#> 44            1  44         -1
#> 45            1  45         -1
#> 46            1  46         -1
#> 47            1  47         -1
#> 48            1  48         -1
#> 49            1  49         -1
#> 50            1  50         -1